requirements.txt
anyio==4.13.0
certifi==2026.4.22
charset-normalizer==3.4.7
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
markdown-it-py==4.0.0
mdurl==0.1.2
psutil==7.2.2
Pygments==2.20.0
python-telegram-bot==22.7
requests==2.33.1
rich==15.0.0
typing_extensions==4.15.0
urllib3==2.6.3
